fix: refactor color, material and pattern logic to utils
